Scehema: silently convert `$table->boolean()` into `INTEGER` on all RDBMS

+++ system/database/schema/grammars/postgres.php
@@ -350,7 +350,7 @@ class Postgres extends Grammar
      */
     protected function type_boolean(Magic $column)
     {
-        return 'SMALLINT';
+        return 'INT';
     }
 
     /**
